AAI/LibriSpeech-SI/requirements.txt
matplotlib~=3.5.2
torchsummary~=1.5.1
torchaudio~=0.13.0
librosa~=0.9.2
numpy~=1.21.5
soundfile~=0.11.0
pathlib~=1.0.1
tqdm~=4.64.1
